correcting results file path

+++ evaluation/evaluation_gptj.py
@@ -113,8 +113,8 @@ if not os.path.exists('results'):
 
 file_name = args.model.split('/')[1]
 
-save_result(qa_gender_ds,'results/' + file_name + '-qa_gender.json', BATCH_SIZE / 2)
-save_result(qa_race_ds,'results/' + file_name + '-qa_race.json', BATCH_SIZE / 2)
+save_result(qa_gender_ds,'results/' + file_name + '-qa_gender.json', int(BATCH_SIZE / 2))
+save_result(qa_race_ds,'results/' + file_name + '-qa_race.json', int(BATCH_SIZE / 2))
 save_result(nli_gender_ds,'results/' + file_name + '-nli_gender.json', BATCH_SIZE)
 save_result(nli_race_ds,'results/' + file_name +  '-nli_race.json', BATCH_SIZE)
 save_result(sentiment_gender_ds,'results/' + file_name +'-sentiment_gender.json', BATCH_SIZE)
